Updating binning to iterate over stimulus table.
Instead of binning data by unique stimulus combinations, the
updated function now bins based on the start/end indices provided
in the stimulus table.  This allows the binned data to be merged
with the stimulus table, which should provide a more general data
format that can be used in various analyses.